Crate imxrt_usbd

A USB driver for i.MX RT processors

imxrt-usbd provides a usb-device USB bus implementation, allowing you to add USB device features to your embedded Rust program. See each module for usage and examples.

To interface the library, you must define a safe implementation of Peripherals. See the trait documentation for more information.

General guidance

The driver does not configure any of the CCM or CCM_ANALOG registers. You are responsible for configuring these peripherals for proper USB functionality. See the imxrt-usbd hardware examples to see different ways of configuring PLLs and clocks.

Modules

gpt

USB general purpose timers.

Structs

BusAdapter

A full- and high-speed UsbBus implementation

EndpointMemory

Memory for endpoint I/O.

EndpointState

Driver state associated with endpoints.

Enums

Speed

USB low / full / high speed setting.

Constants

MAX_ENDPOINTS

The maximum supported number of endpoints.

Traits

Peripherals

A type that owns all USB register blocks