Crate imxrt_usbd

A USB driver for i.MX RT processors

imxrt-usbd provides a usb-device USB bus implementation, allowing you to add USB device features to your embedded Rust program. See each module for usage and examples.

To interface the library, you must define a safe implementation of Peripherals. See the trait documentation for more information.

General guidance

The driver does not configure any of the CCM or CCM_ANALOG registers. You are responsible for configuring these peripherals for proper USB functionality. See the imxrt-usbd hardware examples to see different ways of configuring PLLs and clocks.

Modules

full_speed

A full-speed i.MX RT USB driver, supporting the usb-device ecosystem

gpt

USB general purpose timers.

Enums

Instance

USB instance identifiers

Traits

Peripherals

A type that owns all USB register blocks