Struct imxrt_usbd::full_speed::BusAdapter

pub struct BusAdapter { /* fields omitted */ }

A UsbBus implementation

The BusAdapter adapts the RAL instances, and exposes a UsbBus implementation.

Requirements

The driver assumes that you've prepared all USB clocks (CCM clock gates, CCM analog PLLs). You may use the imxrt-ral APIs to configure USB clocks.

When you build your final usb-device, you must set the endpoint 0 max packet size to 64 bytes. See UsbDeviceBuilder::max_packet_size_0 for more information. Failure to increase the control endpoint max packet size will result in a USB device that cannot communicate with the host.

Additionally, before polling for USB class traffic, you must call configure() after your device has been configured. This can be accomplished by polling the USB device and checking its state until it's been configured. Once configured, use UsbDevice::bus() to access the i.MX RT BusAdapter, and call configure(). You should only do this once. After that, you may poll for class traffic.

Example

This example shows you how to create a BusAdapter, build a simple USB device, and prepare the device for class traffic.

Note that this example does not demonstrate USB class allocation or polling. See your USB class' documentation for details. This example also skips the clock initialization.

use imxrt_usbd::full_speed::BusAdapter;
use imxrt_ral::{usb, usbphy};

static mut ENDPOINT_MEMORY: [u8; 1024] = [0; 1024];

// TODO initialize clocks...

let bus_adapter = BusAdapter::new(
    usb::USB1::take().unwrap(),
    usbphy::USBPHY1::take().unwrap(),
    unsafe { &mut ENDPOINT_MEMORY }
);

// Create the USB device...
use usb_device::prelude::*;
let bus_allocator = usb_device::bus::UsbBusAllocator::new(bus_adapter);
let mut device = UsbDeviceBuilder::new(&bus_allocator, UsbVidPid(0x5824, 0x27dd))
    .product("imxrt-usbd")
    .max_packet_size_0(64) // <---- Set the control OUT/IN max packet size to 64
    // Other builder methods...
    .build();

// Poll until configured...
loop {
    if device.poll(&mut []) {
        let state = device.state();
        if state == usb_device::device::UsbDeviceState::Configured {
            break;
        }
    }
}

// Configure the bus
device.bus().configure();

// Ready for class traffic!

Implementations

impl BusAdapter

pub fn new(usb: Instance, phy: Instance, buffer: &'static mut [u8]) -> Self

Create a USB bus adapter

The two USB instances come from the imxrt-ral register access layer. When this function returns, the BusAdapter has initialized the PHY and USB core registers.

You must also provide a region of memory that will used for endpoint I/O. The memory region will be partitioned for the endpoints, based on their requirements.

You must ensure that no one else is using the endpoint memory!

Panics

Panics if the USB instances are mismatched (a USB1 instance with a USBPHY2 instance).

pub fn set_interrupts(&self, interrupts: bool)

Enable (true) or disable (false) interrupts for this USB peripheral

pub fn configure(&self)

Apply device configurations, and perform other post-configuration actions

You must invoke this once, and only after your device has been configured. See the top-level example for how this could be achieved.