Module imxrt_usbd::gpt

USB general purpose timers.

Each USB OTG peripheral has two general purpose timers (GPT). You can access GPTs through your USB driver.

Example

This example shows how to access a GPT through the BusAdapter API. The example skips the bus adapter and USB device setup in order to focus on the GPT API. See the bus adapter documentation for more information.

use imxrt_usbd::BusAdapter;
use imxrt_usbd::gpt;

let bus_adapter = BusAdapter::new(
    // ...
);

// Prepare a GPT before creating a USB device;
bus_adapter.gpt_mut(gpt::Instance::Gpt0, |gpt| {
    gpt.stop(); // Stop the timer, just in case it's already running...
    gpt.clear_elapsed(); // Clear any outstanding elapsed flags
    gpt.set_interrupt_enabled(false); // Enable or disable interrupts
    gpt.set_load(75_000); // Elapse after 75ms (75000us)
    gpt.set_mode(gpt::Mode::Repeat); // Repeat the timer after it elapses
    gpt.reset(); // Load the value into the counter
});
// The timer isn't running until you call run()...

let bus_allocator = usb_device::bus::UsbBusAllocator::new(bus_adapter);

let mut device = UsbDeviceBuilder::new(&bus_allocator, UsbVidPid(0x5824, 0x27dd))
    .product("imxrt-usbd")
    .build();

// You can still access the timer through the bus() method on
// the USB device.
device.bus().gpt_mut(gpt::Instance::Gpt0, |gpt| gpt.run()); // Timer running!

loop {
    device.bus().gpt_mut(gpt::Instance::Gpt0, |gpt| {
        if gpt.is_elapsed() {
            gpt.clear_elapsed();
            // Timer elapsed!

            // If your mode is Mode::OneShot, you will need
            // to call reset() to re-enable the timer. You also
            // need to call reset() whenever you change the timer
            // load value.
        }
    });
}

Structs

Gpt

General purpose timer (GPT).

Enums

Instance

GPT instance identifiers.

Mode

GPT timer mode.