Struct imxrt_usbd::gpt::Gpt

pub struct Gpt<'a> { /* private fields */ }

General purpose timer (GPT).

USB GPTs have a 1us resolution. The counter is 24 bits wide. GPTs can generate USB interrupts that are independent of USB protocol interrupts. This lets you add additional, time-driven logic into your USB ISR and driver state machine.

See the module-level documentation for an example.

Implementations

impl<'a> Gpt<'a>

pub fn instance(&self) -> Instance

Returns the GPT instance identifier.

pub fn run(&mut self)

Run the GPT timer.

Run will start counting down the timer. Use stop() to cancel a running timer.

pub fn is_running(&self) -> bool

Indicates if the timer is running (true) or stopped (false).

pub fn stop(&mut self)

Stop the timer.

pub fn reset(&mut self)

Reset the timer.

reset loads the counter value. It does not stop a running counter.

pub fn set_mode(&mut self, mode: Mode)

Set the timer mode.

pub fn mode(&self) -> Mode

Returns the timer mode.

pub fn set_load(&mut self, us: u32)

Set the counter load value.

us is the number of microseconds to count. us will saturate at a 24-bit value (0xFFFFFF, or 16.777215 seconds). A value of 0 will result in a 1us delay.

Note that the load count value is not loaded until the next call to reset() (one shot mode) or until after the timer elapses (repeat mode).

pub fn load(&self) -> u32

Returns the counter load value.

pub fn is_elapsed(&self) -> bool

Indicates if the timer has elapsed.

If the timer has elapsed, you should clear the elapsed flag with clear_elapsed().

pub fn clear_elapsed(&mut self)

Clear the flag that indicates the timer has elapsed.

pub fn set_interrupt_enabled(&mut self, enable: bool)

Enable or disable interrupt generation when the timer elapses.

If enabled (true), an elapsed GPT will generate an interrupt. This happens regardless of the USB interrupt enable state.

pub fn is_interrupt_enabled(&self) -> bool

Indicates if interrupt generation is enabled.